Added a TileSet class to keep track of tiles and their attributes. Also edited related files to allow them to incorporate TileSets.
How it works:
 TileSet uses two std::map's to keep track of images and tiles. The first map pairs the image path/filename to a shared pointer of the image's SDL_Texture. The second map pairs the tile name with the actual tile.

The tilename in the map is defined as `<image file name> - <tile name>`. So if forest.png has a tile named 'Grass - Plain', the tile will be queryable by searching map with 'forest - Grass - Plain'.

With a TileSet, you can query for individual parts of the Tile.
mTileSet->Texture("forest - Grass - Plain") will return a pointer to the SDL_Texture associated with the appropriate image.
mTileSet->Clip("forest - Grass - Plain") will return the Recti for the appropriate tile.
mTileSet->Solid('forest - Grass - Plain") will return the bool value for whether or not the tile is solid.

TileSet is loaded in editor & game states, and passed into the map and tilebar classes, so those classes no longer need to load in any images for tiles to function.

To disambiguate, image refers to an actual image, not the Image class. The Image class isn't used anywhere in this new code.