Set up an accessor function, but im unable to get
through it with print(r.x) for ex. it instead prints a function address
of some sort
using it as :accessor("x") works, but i think it should be in __index to
be a getter, but I need the other functions in there as well. Need to
work more on this