trying to setup entity lookup system and registration of entities with manager upon creation
but vs is really being a bitch about it. error messages completely
worthless, WTF even is going on. And for even more fun, the calls work
when done directly from C++, with no Lua involvement. Committing this
code so i can get it on my laptop if i need it later, this is an
unbelievable pain in the ass.