Wrote and tested C/P/A for TimerLib
As a bonus, copying data over is now nearly instant on my machine. Timer
is ticked up to 5 in main menu then passed to quitbutton which prints
the ticks, which are typically 5, sometimes 6. So an elapsed time of
<=1ms. quitbutton then ticks it  up til 10 as I wanted to make sure it
was still timing. This is a nice bump in speed from using the old system
which would take 7-10ms on my desktop (~25-30 I think on my laptop)