Migrated math class to use weak_ptr for camera instead of raw
Also, a weak_ptr is made automatically when passing a shared_ptr to a
function expecting a weak_ptr, the appropriate weak_ptr is created
implicitly. Should remove all unnecessary weak_ptr creations that
existed b/c I didn't know this worked heh. If any of this is still
around

In addition, this commit marks the breaking of the luabind camera
library in regards to using it for passing to the Math library. The code
from this point until the completion of the Lua libraries is now
unusable due to incompatabilities between Lua C API libs and LuaBind