my primitive types and ptr types have some conflicts with how they need to be treated in order
to be able to manipulate the actual object. Ie, the primitive types have
to be used via a pointer, but the other types are already a pointer so
they're fine. Something more clever needs to be thought of to resolve
this + the conflicts with the regular primitives.

Also now i get a jsoncpp assert failure? wtf?