LuaParam and its child classes are gone
Also due to some issues with the location of defining the metatable name
for DataLib specializations it had to be moved into DataLib.
Alternatively luacfunctioninterface could include all the luac libs but
it would've resulted in the same inclusion issue. So due to this issue
LuaScript is now a shared_ptr member of entity, becuase i had to forward
declare the class since it started having circular inclusion of entity
<-> luascript