Retooled Maps (Save, Load, Draw)
I've reworked the way we Save, Load, and Draw Maps.

Save -
Maps are in their own folder in /res/maps.
They consist of 'rows', 'columns' and 'tiles' attributes.
-tiles are name/index pairs where each name has a list of indicies for
where the tile should be drawn based on the # of rows and columns.

Load
Maps are loaded through the game state's 'Map' attribute. This should
link to a map in the /res/maps directory.

Draw
Maps now are only created once. On the first click we create the map
texture and save it into memory. All subsequent draws use the
prerendered image. Calling Map->RebuildMap() will force a redraw.